src/api/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asyncio-redis==0.16.0
asyncpg==0.29.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
llvmlite==0.43.0
lxml==5.2.2
LxmlSoup==1.5.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numba==0.60.0
numpy==2.0.0
orjson==3.10.5
pydantic==2.7.4
pydantic_core==2.18.4
pydentic==0.0.1.dev3
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
python-stdnum==1.20
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
termcolor==2.4.0
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
